refactor: merge auth routes and refactor root routes specs ds-48
- update some transitions.
- remove app routes constants, since we navigate via route types.
- add missing appbar on post story.
