--- lib/interfaces/libs/constants/route_paths/app_route_paths_constant.dart
@@ -1,5 +1,0 @@
-abstract final class AppRoutePaths {
-  static const auth = "/auth";
-  static const stories = "/stories";
-  static const profile = "/profile";
-}
