--- lib/interfaces/libs/constants/route_paths/auth_route_paths_constant.dart
@@ -1,7 +1,0 @@
-import "app_route_paths_constant.dart";
-
-/// Sub paths of [AppRoutePaths.auth].
-abstract final class AuthRoutePaths {
-  static const signIn = "/sign-in";
-  static const signUp = "/sign-up";
-}
