--- lib/interfaces/libs/constants/route_paths/profile_route_paths_constant.dart
@@ -1,6 +1,0 @@
-import "app_route_paths_constant.dart";
-
-/// Sub paths of [AppRoutePaths.profile].
-abstract final class ProfileRoutePaths {
-  static const root = "";
-}
