--- lib/interfaces/libs/constants/route_paths/stories_route_paths_constant.dart
@@ -1,8 +1,0 @@
-import "app_route_paths_constant.dart";
-
-/// Sub paths of [AppRoutePaths.stories].
-abstract final class StoriesRoutePaths {
-  static const root = "";
-  static const post = "post";
-  static const view$storyId = "view/:storyId";
-}
