+++ lib/interfaces/modules/_screens.dart
@@ -1,0 +1,8 @@
+/// Screens of root routes, partitioned from the `routes.dart` file
+/// for easy maintenance in the future.
+///
+/// Sub-routes screens will still defined on its own route file.
+library;
+
+export "_screens/auth_route_screen.dart";
+export "_screens/root_shell_route_screen.dart";
