--- lib/interfaces/modules/routes/auth.dart
@@ -1,2 +1,0 @@
-export "auth/sign_in_route.dart";
-export "auth/sign_up_route.dart";
